przeorganizacja projektu z usunieciem pliku Node.java
